+++ src/program/KeyMappingXcb.h
@@ -66,7 +66,7 @@ class KeyMappingXcb : public KeyMapping {
          * - Get the mouse state
          * - Warp mouse pointer if needed
          */
-        void buildAllInputs(AllInputs& ai, uint32_t window, SharedConfig& sc, bool mouse_warp, int mouse_wheel);
+        void buildAllInputs(AllInputs& ai, uint32_t window, SharedConfig& sc, bool mouse_warp);
 
     private:
         /* Connection to the X11 server */